Update project.ipynb
After the last commit, EDA section was fully added. In addition to that, train_test_split
was done and Logistic Regression model was applied to the dataset. Evaluation was also done for
both scaled as well as unscaled data. Now the only sections of the project left is finishing the
markdown in the project file itself, updating the readme accordingly and finally deploying it to the
internet using frontend, backend and api.